+++ app/config/routes.js
@@ -1,5 +1,6 @@
 import CanDefineMap from 'can-define/map/map';
-import route from 'can-route/can-route';
+import route from 'can-route';
+import RoutePushstate from 'can-route-pushstate';
 
 const AppViewModel = CanDefineMap.extend({
   controller: 'string',
@@ -7,33 +8,34 @@ const AppViewModel = CanDefineMap.extend({
 });
 const appState = new AppViewModel();
 route.data = appState;
-
-// Default route
-route('', {controller: 'Password', action: 'index'});
-
-route.register('/debug', {controller: 'Debug', action: 'index'});
+route.urlData = new RoutePushstate();
 
 // Groups routes
-route.register('/groups/delete/{id}', {controller: 'User', action: 'groupDelete'});
-route.register('/groups/edit/{id}', {controller: 'User', action: 'groupEdit'});
-route.register('/groups/view/{id}', {controller: 'User', action: 'groupView'});
-route.register('/groups/view/{id}/membership', {controller: 'User', action: 'groupViewGroupsUsers'});
+route.register('/app/groups/delete/{id}', {controller: 'User', action: 'groupDelete'});
+route.register('/app/groups/edit/{id}', {controller: 'User', action: 'groupEdit'});
+route.register('/app/groups/view/{id}', {controller: 'User', action: 'groupView'});
+route.register('/app/groups/view/{id}/membership', {controller: 'User', action: 'groupViewMembership'});
 
 // Passwords routes
-route.register('/passwords', {controller: 'Password', action: 'index'});
-route.register('/passwords/edit/{id}', {controller: 'Password', action: 'edit'});
-route.register('/passwords/view/{id}', {controller: 'Password', action: 'view'});
+route.register('/app/passwords', {controller: 'Password', action: 'index'});
+route.register('/app/passwords/edit/{id}', {controller: 'Password', action: 'edit'});
+route.register('/app/passwords/view/{id}', {controller: 'Password', action: 'view'});
+route.register('/app/passwords/view/{id}/comments', {controller: 'Password', action: 'commentsView'});
 
 // Users routes
-route.register('/users', {controller: 'User', action: 'index'});
-route.register('/users/add', {controller: 'User', action: 'add'});
-route.register('/users/edit/{id}', {controller: 'User', action: 'edit'});
-route.register('/users/delete/{id}', {controller: 'User', action: 'delete'});
-route.register('/users/view/{id}', {controller: 'User', action: 'view'});
+route.register('/app/users', {controller: 'User', action: 'index'});
+route.register('/app/users/add', {controller: 'User', action: 'add'});
+route.register('/app/users/edit/{id}', {controller: 'User', action: 'edit'});
+route.register('/app/users/delete/{id}', {controller: 'User', action: 'delete'});
+route.register('/app/users/view/{id}', {controller: 'User', action: 'view'});
 
 // User settings routes
-route.register('/settings', {controller: 'Settings', action: 'profile'});
-route.register('/settings/keys', {controller: 'Settings', action: 'keys'});
-route.register('/settings/profile', {controller: 'Settings', action: 'profile'});
-route.register('/settings/theme', {controller: 'Settings', action: 'theme'});
+route.register('/app/settings/keys', {controller: 'Settings', action: 'keys'});
+route.register('/app/settings/profile', {controller: 'Settings', action: 'profile'});
+route.register('/app/settings/theme', {controller: 'Settings', action: 'theme'});
+route.register('/app/settings', {controller: 'Settings', action: 'profile'});
+
+// Default route
+route.register('', {controller: 'Password', action: 'index'});
+
 route.start();